Moved tests to separated file + added tests for unhandled exceptions
