+++ Tests/Runtime/Client.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: af9b82bb32c57d541b6c9b92a8abfdbb
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
