+++ Tests/Runtime/Client/BacktraceClientDisableTests.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 46d57d41fbb8721488d936e8abdea818
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
